backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
feedparser==6.0.10
textblob==0.17.1
vaderSentiment==3.3.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
eventlet==0.33.3
APScheduler==3.10.4
lxml==4.9.3
nltk==3.8.1
anthropic==0.25.0